Tile Class, Selection and Tile Swaps, debug stream
Added the Tile class, an extension of QGraphicsPixmapItem. It is used to override some functions that select and move the tiles, so that it works the way I want.
I also moved tileHH (highlight) variables into the view class. This is so the tiles may access their values by calling scene->view functions (psuedocode)..

I added snapping algorithms that seem to be working well :)